git-svn-id: http://svn.code.sf.net/p/distributedcl/code/trunk@174 cb344d40-e590-48b6-843f-7a98a4fc1ff3

+++ src/libdistributedcl/message/msg_program.cpp
@@ -168,7 +168,7 @@ void dcl_message< msgGetProgramBuildInfo >::create_response( uint8_t* payload_pt
     }
     else if( build_info_ == CL_PROGRAM_BUILD_LOG )
     {
-        *response_ptr = host_to_network( build_log_.length() );
+        *response_ptr = host_to_network( static_cast<uint32_t>( build_log_.length() ) );
 
         memcpy( response_ptr+1, build_log_.data(), build_log_.length() );
     }
